apps/cms: add captcha to EmailForm
