+++ meinberlin/apps/cms/models/__init__.py
@@ -1,3 +1,4 @@
+from .email_form import *  # noqa
 from .images import *  # noqa
 from .navigation_menues import *  # noqa
 from .pages import *  # noqa
